requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.36
alembic==1.13.3
redis==5.1.1
rq==1.16.2
# google-generativeai==0.8.3  # Replaced by OpenRouter API via httpx
python-ulid==3.0.0
python-dotenv==1.0.1
pillow==10.4.0
httpx==0.27.2
PyJWT==2.9.0
pywebpush==2.0.0
sentry-sdk[fastapi]==2.19.2
slowapi==0.1.9
pytest==8.3.3
pytest-asyncio==0.24.0
pyyaml==6.0.2
